+++ src/node.rs
@@ -2,20 +2,17 @@ use std::{
     collections::HashMap,
     sync::{
         atomic::{AtomicUsize, Ordering},
-        Arc, Mutex, OnceLock,
+        Arc, Mutex,
     },
 };
 
-use crate::message::{Body, Message};
+use crate::{
+    message::{Body, Message},
+    node2::Node2,
+};
 
 type NodeName = String;
 type ClientName = String;
-type NodeOrClientName = String;
-
-struct MessageWithResponder {
-    msg: Message,
-    responder: Option<tokio::sync::oneshot::Sender<Message>>,
-}
 
 #[derive(Debug, Clone)]
 struct ClientTableEntry {
@@ -30,10 +27,6 @@ enum NodeStatus {
 }
 
 pub struct Node {
-    unacked: Mutex<HashMap<usize, tokio::sync::oneshot::Sender<Message>>>,
-    stdout_tx: OnceLock<tokio::sync::mpsc::Sender<MessageWithResponder>>,
-    next_msg_id: AtomicUsize,
-
     // vsr
     op_log: Mutex<Vec<Message>>,
     op_number: AtomicUsize,
@@ -44,15 +37,12 @@ pub struct Node {
     configuration: Mutex<Vec<NodeName>>, // All nodes in the cluster
     replica_number: AtomicUsize,
     client_table: Mutex<HashMap<ClientName, ClientTableEntry>>,
+    node2: Arc<Node2>,
 }
 
 impl Node {
     pub fn new() -> Self {
         Node {
-            unacked: Default::default(),
-            stdout_tx: OnceLock::new(),
-            next_msg_id: AtomicUsize::new(0),
-
             op_log: Mutex::new(Vec::new()),
             op_number: AtomicUsize::new(0),
             commit_number: AtomicUsize::new(0),
@@ -62,26 +52,17 @@ impl Node {
             configuration: Default::default(),
             replica_number: AtomicUsize::new(0),
             client_table: Default::default(),
+            node2: Arc::new(Node2::new()),
         }
     }
 
     pub async fn run(self: Arc<Self>) {
-        let mut stdin_rx = self.clone().spawn_stdin_task().await;
-        self.clone().spawn_stdout_task().await;
-
-        loop {
-            tokio::select! {
-                Some(msg) = stdin_rx.recv() => {
-                    if let Some(responder) = self.unacked.lock().unwrap().remove(&msg.body.msg_id()) {
-                        responder.send(msg).unwrap();
-                    } else {
-                        tokio::spawn({
-                            let node = self.clone();
-                            async move {node.handle(msg).await}
-                        });
-                    }
-                }
-            }
+        let mut rx = self.node2.clone().run().await;
+        while let Some(msg) = rx.recv().await {
+            tokio::spawn({
+                let node = self.clone();
+                async move { node.handle(msg).await }
+            });
         }
     }
 
@@ -111,16 +92,18 @@ impl Node {
                         configuration_guard.first().unwrap().clone();
                 }
 
-                let msg_id = self.reserve_next_msg_id();
-                self.send(
-                    msg.src,
-                    Body::InitOk {
-                        msg_id,
-                        in_reply_to,
-                    },
-                    None,
-                )
-                .await
+                let msg_id = self.node2.reserve_next_msg_id();
+                self.node2
+                    .clone()
+                    .send(
+                        msg.src,
+                        Body::InitOk {
+                            msg_id,
+                            in_reply_to,
+                        },
+                        None,
+                    )
+                    .await
             }
             Body::Write { msg_id, .. } | Body::Read { msg_id, .. } | Body::Cas { msg_id, .. } => {
                 let mut response = None;
@@ -133,20 +116,27 @@ impl Node {
                     Some(ClientTableEntry {
                         op,
                         response: Some(cached_response),
-                    }) if msg == op => self.send(msg.src, cached_response.body, None).await,
+                    }) if msg == op => {
+                        self.node2
+                            .clone()
+                            .send(msg.src, cached_response.body, None)
+                            .await
+                    }
                     Some(ClientTableEntry { op, .. }) if msg.body.msg_id() < op.body.msg_id() => {
                         let error_text =
                             String::from("Normal Protocol: received a stale request number");
-                        self.send(
-                            msg.src,
-                            Body::Error {
-                                in_reply_to: msg.body.msg_id(),
-                                code: crate::message::ErrorCode::Abort,
-                                text: error_text,
-                            },
-                            None,
-                        )
-                        .await
+                        self.node2
+                            .clone()
+                            .send(
+                                msg.src,
+                                Body::Error {
+                                    in_reply_to: msg.body.msg_id(),
+                                    code: crate::message::ErrorCode::Abort,
+                                    text: error_text,
+                                },
+                                None,
+                            )
+                            .await
                     }
                     // request is valid and is not cached in the client table
                     _ => self.handle_client_request(msg).await,
@@ -169,133 +159,5 @@ impl Node {
         //                    n: op number assigned to this request
         //                    k: commit number
         //
-    }
-
-    async fn broadcast(
-        self: Arc<Self>,
-        body: Body,
-        responder: Option<tokio::sync::mpsc::Sender<Message>>,
-    ) {
-        let mut receiver_tasks = tokio::task::JoinSet::<Message>::new();
-
-        let mut other_node_ids = vec![];
-        {
-            let configuration_guard = self.configuration.lock().unwrap();
-
-            let my_id = configuration_guard
-                .get(self.replica_number.load(Ordering::SeqCst))
-                .unwrap();
-
-            other_node_ids = configuration_guard
-                .iter()
-                .filter(|id| *id != my_id)
-                .cloned()
-                .collect();
-        }
-
-        for destination in other_node_ids {
-            let (tx, rx) = tokio::sync::oneshot::channel::<Message>();
-            receiver_tasks.spawn(async move {
-                rx.await
-                    .expect("should be able to recv on one of the broadcast responses")
-            });
-
-            let new_msg_id = self.clone().reserve_next_msg_id();
-
-            let mut body = body.clone();
-            body.set_msg_id(new_msg_id);
-            self.clone().send(destination, body, Some(tx)).await;
-        }
-
-        tokio::spawn(async move {
-            while let Some(response_result) = receiver_tasks.join_next().await {
-                let response_message =
-                    response_result.expect("should be able to recv response during broadcast");
-                if let Some(ref responder) = responder {
-                    responder
-                        .send(response_message)
-                        .await
-                        .expect("should be able to return response message from broadcast()");
-                }
-            }
-        });
-    }
-
-    async fn send(
-        self: Arc<Self>,
-        dest: NodeOrClientName,
-        body: Body,
-        responder: Option<tokio::sync::oneshot::Sender<Message>>,
-    ) {
-        let stdout_tx = self.stdout_tx.get().unwrap();
-
-        let mut my_id = String::new();
-        {
-            let replica_number = self.replica_number.load(Ordering::SeqCst);
-            let configuration = self.configuration.lock().unwrap();
-            my_id = configuration.get(replica_number).unwrap().clone();
-        }
-
-        stdout_tx
-            .send(MessageWithResponder {
-                msg: Message {
-                    src: my_id,
-                    dest,
-                    body,
-                },
-                responder,
-            })
-            .await
-            .unwrap();
-    }
-
-    async fn spawn_stdout_task(self: Arc<Self>) {
-        let (stdout_tx, mut stdout_rx) = tokio::sync::mpsc::channel::<MessageWithResponder>(32);
-
-        self.stdout_tx.set(stdout_tx).unwrap();
-
-        tokio::spawn(async move {
-            while let Some(MessageWithResponder { msg, responder }) = stdout_rx.recv().await {
-                println!(
-                    "{}",
-                    serde_json::to_string(&msg)
-                        .expect("msg being sent to STDOUT should be serializable to JSON")
-                );
-                tracing::debug!("sent msg {:?}", &msg);
-
-                if let Some(responder) = responder {
-                    self.unacked
-                        .lock()
-                        .unwrap()
-                        .insert(msg.body.msg_id(), responder);
-                }
-            }
-        });
-    }
-
-    async fn spawn_stdin_task(self: Arc<Self>) -> tokio::sync::mpsc::Receiver<Message> {
-        let (stdin_tx, stdin_rx) = tokio::sync::mpsc::channel::<Message>(32);
-        tokio::spawn(async move {
-            let mut input = String::new();
-            let mut is_reading_stdin = true;
-            while is_reading_stdin {
-                if let Err(e) = std::io::stdin().read_line(&mut input) {
-                    println!("readline error: {e}");
-                    is_reading_stdin = false;
-                }
-
-                let json_msg = serde_json::from_str(&input)
-                    .expect(&format!("should take a JSON message. Got {:?}", input));
-                tracing::debug!("received json msg: {:?}", json_msg);
-
-                stdin_tx.send(json_msg).await.unwrap();
-                input.clear();
-            }
-        });
-        stdin_rx
-    }
-
-    fn reserve_next_msg_id(&self) -> usize {
-        self.next_msg_id.fetch_add(1, Ordering::SeqCst)
     }
 }
