+++ src/node2.rs
@@ -1,0 +1,197 @@
+use std::{
+    collections::HashMap,
+    sync::{
+        atomic::{AtomicUsize, Ordering},
+        Arc, Mutex, OnceLock,
+    },
+};
+
+use crate::message::{Body, Message};
+
+type NodeOrClientName = String;
+
+struct MessageWithResponder {
+    msg: Message,
+    responder: Option<tokio::sync::oneshot::Sender<Message>>,
+}
+
+pub struct Node2 {
+    pub my_id: OnceLock<String>,
+    pub other_node_ids: OnceLock<Vec<String>>,
+    pub unacked: Arc<Mutex<HashMap<usize, tokio::sync::oneshot::Sender<Message>>>>,
+    pub stdout_tx: OnceLock<tokio::sync::mpsc::Sender<MessageWithResponder>>,
+    pub next_msg_id: AtomicUsize,
+}
+
+impl Node2 {
+    pub fn new() -> Self {
+        Self {
+            unacked: Default::default(),
+            stdout_tx: OnceLock::new(),
+            next_msg_id: AtomicUsize::new(0),
+            my_id: OnceLock::new(),
+            other_node_ids: OnceLock::new(),
+        }
+    }
+
+    pub async fn send(
+        self: Arc<Self>,
+        dest: NodeOrClientName,
+        body: Body,
+        responder: Option<tokio::sync::oneshot::Sender<Message>>,
+    ) {
+        let stdout_tx = self.stdout_tx.get().unwrap();
+
+        stdout_tx
+            .send(MessageWithResponder {
+                msg: Message {
+                    src: self.my_id.get().unwrap().into(),
+                    dest,
+                    body,
+                },
+                responder,
+            })
+            .await
+            .unwrap();
+    }
+
+    pub async fn broadcast(
+        self: Arc<Self>,
+        body: Body,
+        responder: Option<tokio::sync::mpsc::Sender<Message>>,
+    ) {
+        let mut receiver_tasks = tokio::task::JoinSet::<Message>::new();
+
+        for destination in self.other_node_ids.get().unwrap().clone() {
+            let (tx, rx) = tokio::sync::oneshot::channel::<Message>();
+            receiver_tasks.spawn(async move {
+                rx.await
+                    .expect("should be able to recv on one of the broadcast responses")
+            });
+
+            let new_msg_id = self.clone().reserve_next_msg_id();
+
+            let mut body = body.clone();
+            body.set_msg_id(new_msg_id);
+            self.clone().send(destination, body, Some(tx)).await;
+        }
+
+        tokio::spawn(async move {
+            while let Some(response_result) = receiver_tasks.join_next().await {
+                let response_message =
+                    response_result.expect("should be able to recv response during broadcast");
+                if let Some(ref responder) = responder {
+                    responder
+                        .send(response_message)
+                        .await
+                        .expect("should be able to return response message from broadcast()");
+                }
+            }
+        });
+    }
+
+    pub async fn run(self: Arc<Self>) -> tokio::sync::mpsc::Receiver<Message> {
+        let mut stdin_rx = self.clone().spawn_stdin_task().await;
+        self.clone().spawn_stdout_task().await;
+
+        let (tx, rx) = tokio::sync::mpsc::channel(32);
+
+        tokio::spawn(async move {
+            while let Some(msg) = stdin_rx.recv().await {
+                // std::mem::drop(unacked);
+
+                let mut r: Option<tokio::sync::oneshot::Sender<Message>> = None;
+                {
+                    let mut unacked = self.unacked.lock().unwrap();
+                    r = unacked.remove(&msg.body.msg_id());
+                }
+
+                if let Some(responder) = r {
+                    responder.send(msg).unwrap();
+                } else {
+                    tx.send(msg).await.unwrap();
+                }
+            }
+        });
+
+        rx
+    }
+
+    // let s = self.clone();
+    // tokio::spawn(async move {
+    //     while let Some(msg) = stdin_rx.recv().await {
+    //         if let Some(responder) =
+    //             s.unacked.clone().lock().unwrap().remove(&msg.body.msg_id())
+    //         {
+    //             responder.send(msg).unwrap();
+    //         } else {
+    //             tx.send(msg).await.unwrap();
+    //         }
+    //     }
+    // });
+
+    pub async fn spawn_stdout_task(self: Arc<Self>) {
+        let (stdout_tx, mut stdout_rx) = tokio::sync::mpsc::channel::<MessageWithResponder>(32);
+
+        self.stdout_tx.set(stdout_tx).unwrap();
+
+        tokio::spawn(async move {
+            while let Some(MessageWithResponder { msg, responder }) = stdout_rx.recv().await {
+                println!(
+                    "{}",
+                    serde_json::to_string(&msg)
+                        .expect("msg being sent to STDOUT should be serializable to JSON")
+                );
+                tracing::debug!("sent msg {:?}", &msg);
+
+                if let Some(responder) = responder {
+                    self.unacked
+                        .lock()
+                        .unwrap()
+                        .insert(msg.body.msg_id(), responder);
+                }
+            }
+        });
+    }
+
+    pub async fn spawn_stdin_task(self: Arc<Self>) -> tokio::sync::mpsc::Receiver<Message> {
+        let (stdin_tx, stdin_rx) = tokio::sync::mpsc::channel::<Message>(32);
+        tokio::spawn(async move {
+            let mut input = String::new();
+            let mut is_reading_stdin = true;
+            while is_reading_stdin {
+                if let Err(e) = std::io::stdin().read_line(&mut input) {
+                    println!("readline error: {e}");
+                    is_reading_stdin = false;
+                }
+
+                let json_msg: Message = serde_json::from_str(&input)
+                    .expect(&format!("should take a JSON message. Got {:?}", input));
+                tracing::debug!("received json msg: {:?}", json_msg);
+
+                if let Body::Init {
+                    node_id, node_ids, ..
+                } = &json_msg.body
+                {
+                    self.my_id.set(node_id.into());
+
+                    self.other_node_ids.set(
+                        node_ids
+                            .iter()
+                            .filter(|id| *id != node_id)
+                            .cloned()
+                            .collect(),
+                    );
+                }
+
+                stdin_tx.send(json_msg).await.unwrap();
+                input.clear();
+            }
+        });
+        stdin_rx
+    }
+
+    pub fn reserve_next_msg_id(&self) -> usize {
+        self.next_msg_id.fetch_add(1, Ordering::SeqCst)
+    }
+}
